// com.ptoceti.osgi.ihm/src/main/webapp/ihm/src/app/api/model/configurationWrapper.ts
/**
 * Ptoceti Rest Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import {ConfigurationPropertyEntry} from './configurationPropertyEntry';


export interface ConfigurationWrapper {
  bundleLocation?: string;
  factoryPid?: string;
  pid?: string;
  properties?: Array<ConfigurationPropertyEntry>;
}